refactor: RewardCalculator를 global util로 이동
#62
