+++ src/main/java/com/explorer/gabom/global/util/RewardCalculator.java
@@ -1,4 +1,6 @@
-package com.explorer.gabom.domain.exploration.vo;
+package com.explorer.gabom.global.util;
+
+import com.explorer.gabom.domain.exploration.vo.RewardRange;
 
 public class RewardCalculator {
